refactor(server): split eval-runner.ts into lib/eval-runners/ (PR 4/4 tech-debt)
Final PR of the 2026-06-12 tech-debt pass. PRs A (#318), B (#319), C
(#320) already landed.

lib/eval-runner.ts was 1273 lines, the second-largest single file in
the server after pre-PR-B cron.ts. It mixed three independent concerns:

  1. LLM-as-judge runner (callJudge, pool, generateForItem, dispatcher)
  2. Deterministic runners (runRegex, runJsonSchema, runSimpleEvalRun)
  3. Shared response-text extraction

Extracted (3 new files):
- lib/eval-runners/shared.ts        (48 lines)
    MAX_RESPONSE_CHARS + extractResponseText. Pure, no I/O.
- lib/eval-runners/judge-prompt.ts (210 lines)
    JudgeConfig + TypedScoreConfig types, buildJudgePrompt,
    parseJudgeReply. The pure side-effect-free core of the judge path —
    all DB / network / cost logic stays in eval-runner.ts as the
    imperative shell.
- lib/eval-runners/deterministic.ts (216 lines)
    RegexConfig + JsonSchemaConfig + SimpleEvalResult, runRegex,
    runJsonSchema, runSimpleEvalRun. Completely independent of the LLM
    judge path.

lib/eval-runner.ts: 1273 → 882 lines (-391). Still above the 800-line
ceiling but down 31% and now contains only one execution model (LLM
judge + dispatcher), not three. Further reduction is a future PR.

Backward compatibility: all existing import sites are preserved via
re-exports at the top of eval-runner.ts. No call site (api/evals.ts,
the two eval-runner test files) was edited — they all still import from
'../lib/eval-runner.js' and get the same symbols.

Verification:
- pnpm --filter server test: 1022/1022 (unchanged from PR C)
- pnpm --filter server typecheck: clean
- 23 eval-runner unit tests pass against the extracted modules
- 6 eval-runner-code-evaluators tests pass against the extracted
  deterministic runners

This closes the 2026-06-12 tech-debt audit. Summary across all 4 PRs:

  proxy/openai.ts:    245 → 129 lines  (-47%)
  proxy/anthropic.ts: 220 → 125 lines  (-43%)
  proxy/gemini.ts:    283 → 186 lines  (-34%)
  proxy/azure.ts:     251 → 149 lines  (-41%)
  api/cron.ts:       1053 → 383 lines  (-64%)
  lib/eval-runner:   1273 → 882 lines  (-31%)

  console.error sites in hot paths replaced with structured logger
  fmtCost duplicated across 19 files → 3 named helpers, 10 sites migrated

  Test count: 946 → 1022 (+76 across PRs)
